Avalon

+++ Avalon/Avalon.Network/Status.cs
@@ -2,5 +2,5 @@ namespace Avalon.Network;
 
 public class Status : Any
 {
-    public virtual int Index { get; set; }
+    public virtual long Index { get; set; }
 }
